modificado editar peril - medico,paciente mockuops

+++ app/Http/Controllers/perfilController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\User;
 use App\Role;
 use App\Pregunta;
+use Illuminate\Support\Facades\Auth;
 
 class perfilController extends Controller
 {
@@ -20,9 +21,13 @@ class perfilController extends Controller
     {
         $questions = Pregunta::all();
         $roles = Role::all();
+        $rol = $user->roles()->get()->toArray();
+        $isAdmin = Auth::user()->hasRole('admin');
         return view('perfil.showEdit')
         ->with('user', $user)
+        ->with('isAdmin', $isAdmin)
         ->with('roles', $roles)
+        ->with('rol', $rol)
         ->with('questions', $questions);
     }
 
@@ -32,21 +37,31 @@ class perfilController extends Controller
         $userEdit->email = $request->email;
         $userEdit->ci = $request->ci;
         $userEdit->username = $request->username;
-        $avatar = $request->avatar;
-        $userEdit->avatar = $avatar->store('users', 'public');
+        // $avatar = $request->avatar;
+        // $userEdit->avatar = $avatar->store('users', 'public');
 
         $userEdit->save();
 
         $role_entry = Role::find($request->input('role'));
+
         
-        $userEdit->roles()->detach();
-        $userEdit->attachRole($role_entry);
+        if($role_entry){
+            $userEdit->roles()->detach();
+            $userEdit->attachRole($role_entry);
+        }
+        
+        return redirect()->back()->with('message', 'Perfil Actualizado');
+    }
+
+    public function editPreguntas(Request $request){
+
+        $userEdit = User::find($request->user_id);
 
         $userEdit->preguntas()->detach();
         $userEdit->preguntas()->attach(1, ['respuesta' => $request->pregunta1]);
         $userEdit->preguntas()->attach(2, ['respuesta' => $request->pregunta2]);
         $userEdit->preguntas()->attach(3, ['respuesta' => $request->pregunta3]);
 
-        return redirect('/perfil');
+        return redirect()->back()->with('message', 'Preguntas Actualizadas');
     }
 }
